// src/types/ITemplateItem.ts
export interface ITemplateItem {
    _id: string;
    _name: string;
    _parent: string;
    _type: string;
    _props: IProps;
    _proto?: string;
}